fix remaining low effort warnings not caught by dotnet format
