+++ src/Dapr/GameServer.Host/GameServerController.cs
@@ -30,7 +30,7 @@ public class GameServerController : ControllerBase
     /// <summary>
     /// Shuts down the server gracefully.
     /// </summary>
-    /// <returns></returns>
+    /// <returns>A <see cref="ValueTask"/> representing the asynchronous operation.</returns>
     [HttpPost(nameof(IGameServer.ShutdownAsync))]
     public async ValueTask ShutdownAsync()
     {
