+++ src/GameLogic/Views/IConsumeSpecialItemPlugIn.cs
@@ -14,6 +14,6 @@ public interface IConsumeSpecialItemPlugIn : IViewPlugIn
     /// </summary>
     /// <param name="item">The item.</param>
     /// <param name="effectTimeInSeconds">The effect time in seconds.</param>
-    /// <returns></returns>
+    /// <returns>A <see cref="ValueTask"/> representing the asynchronous operation.</returns>
     ValueTask ConsumeSpecialItemAsync(Item item, ushort effectTimeInSeconds);
 }
