+++ src/Web/Shared/ComponentBuilders/DateOnlyFieldBuilder.cs
@@ -1,0 +1,22 @@
+// <copyright file="DateOnlyFieldBuilder.cs" company="MUnique">
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace MUnique.OpenMU.Web.Shared.ComponentBuilders;
+
+using System.Reflection;
+using Microsoft.AspNetCore.Components.Rendering;
+using MUnique.OpenMU.Web.Shared.Components.Form;
+using MUnique.OpenMU.Web.Shared.Services;
+
+/// <summary>
+/// A <see cref="IComponentBuilder"/> for date fields.
+/// </summary>
+public class DateOnlyFieldBuilder : BaseComponentBuilder, IComponentBuilder
+{
+    /// <inheritdoc/>
+    public int BuildComponent(object model, PropertyInfo propertyInfo, RenderTreeBuilder builder, int currentIndex, IChangeNotificationService notificationService) => this.BuildField<DateOnly, DateField>(model, propertyInfo, builder, currentIndex, notificationService);
+
+    /// <inheritdoc/>
+    public bool CanBuildComponent(PropertyInfo propertyInfo) => propertyInfo.PropertyType == typeof(DateOnly);
+}
