+++ src/Web/Shared/ComponentBuilders/DateTimeFieldBuilder.cs
@@ -20,39 +20,3 @@ public class DateTimeFieldBuilder : BaseComponentBuilder, IComponentBuilder
     /// <inheritdoc/>
     public bool CanBuildComponent(PropertyInfo propertyInfo) => propertyInfo.PropertyType == typeof(DateTime);
 }
-
-/// <summary>
-/// A <see cref="IComponentBuilder"/> for date fields.
-/// </summary>
-public class DateOnlyFieldBuilder : BaseComponentBuilder, IComponentBuilder
-{
-    /// <inheritdoc/>
-    public int BuildComponent(object model, PropertyInfo propertyInfo, RenderTreeBuilder builder, int currentIndex, IChangeNotificationService notificationService) => this.BuildField<DateOnly, DateField>(model, propertyInfo, builder, currentIndex, notificationService);
-
-    /// <inheritdoc/>
-    public bool CanBuildComponent(PropertyInfo propertyInfo) => propertyInfo.PropertyType == typeof(DateOnly);
-}
-
-/// <summary>
-/// A <see cref="IComponentBuilder"/> for date fields.
-/// </summary>
-public class TimeOnlyFieldBuilder : BaseComponentBuilder, IComponentBuilder
-{
-    /// <inheritdoc/>
-    public int BuildComponent(object model, PropertyInfo propertyInfo, RenderTreeBuilder builder, int currentIndex, IChangeNotificationService notificationService) => this.BuildField<TimeOnly, TimeField>(model, propertyInfo, builder, currentIndex, notificationService);
-
-    /// <inheritdoc/>
-    public bool CanBuildComponent(PropertyInfo propertyInfo) => propertyInfo.PropertyType == typeof(TimeOnly);
-}
-
-/// <summary>
-/// A <see cref="IComponentBuilder"/> for time span fields.
-/// </summary>
-public class TimeSpanFieldBuilder : BaseComponentBuilder, IComponentBuilder
-{
-    /// <inheritdoc/>
-    public int BuildComponent(object model, PropertyInfo propertyInfo, RenderTreeBuilder builder, int currentIndex, IChangeNotificationService notificationService) => this.BuildField<TimeSpan, TimeSpanField>(model, propertyInfo, builder, currentIndex, notificationService);
-
-    /// <inheritdoc/>
-    public bool CanBuildComponent(PropertyInfo propertyInfo) => propertyInfo.PropertyType == typeof(TimeSpan);
-}
