+++ src/Web/Shared/Components/MapEditor/MapCoordinateService.cs
@@ -12,12 +12,12 @@ using System.Globalization;
 /// </summary>
 public sealed class MapCoordinateService
 {
+    private const int MapSize = 256;
+
     /// <summary>
     /// Gets the base pixel scale factor before zoom is applied.
     /// </summary>
     public static int BaseScale => 3;
-
-    private const int MapSize = 256;
 
     /// <summary>
     /// Converts client mouse coordinates to map tile coordinates.
